update 비디오 룸
