+++ components/LiveRoom/Video.tsx
@@ -1,34 +1,11 @@
 import React, { useEffect, useRef, useState } from 'react';
-import styled from '@emotion/styled';
-
-const Container = styled.div`
-  position: relative;
-  display: inline-block;
-  width: 240px;
-  height: 270px;
-  margin: 5px;
-`;
-
-const VideoContainer = styled.video`
-  width: 240px;
-  height: 240px;
-  background-color: black;
-`;
-
-const UserLabel = styled.p`
-  display: inline-block;
-  position: absolute;
-  top: 230px;
-  left: 0px;
-`;
 
 interface Props {
-  email: string;
   stream: MediaStream;
   muted?: boolean;
 }
 
-const Video = ({ email, stream, muted }: Props) => {
+const Video = ({ stream, muted }: Props) => {
   const ref = useRef<HTMLVideoElement>(null);
   const [isMuted, setIsMuted] = useState<boolean>(false);
 
@@ -37,12 +14,7 @@ const Video = ({ email, stream, muted }: Props) => {
     if (muted) setIsMuted(muted);
   }, [stream, muted]);
 
-  return (
-    <Container>
-      <VideoContainer ref={ref} muted={isMuted} autoPlay />
-      <UserLabel>{email}</UserLabel>
-    </Container>
-  );
+  return <video ref={ref} muted={isMuted} autoPlay />;
 };
 
 export default Video;
