+++ components/LiveRoom/index.tsx
@@ -1,10 +1,11 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useState, useRef, useEffect, useCallback } from 'react';
 import io from 'socket.io-client';
+import { SOCKET_SERVER_URL } from '../../constants/env';
 import styled from '@emotion/styled';
 import VideoDetailInfor from '../Common/VideoDetailInfor';
-import Send from '../../assets/send';
 import Comment from '../Common/Comment';
-import { SOCKET_SERVER_URL } from '../../constants/env';
+import Send from '../../assets/send';
+import Video from './Video';
 
 const pc_config = {
   iceServers: [
@@ -15,128 +16,202 @@ const pc_config = {
 };
 const SERVER_URL = SOCKET_SERVER_URL as string;
 
-export default function LiveBroadcastRoom() {
+export default function LiveRoom() {
   const socketRef = useRef<SocketIOClient.Socket>();
-  const pcRef = useRef<RTCPeerConnection>();
+  const pcsRef = useRef<{ [socketId: string]: RTCPeerConnection }>({});
   const localVideoRef = useRef<HTMLVideoElement>(null);
-  const remoteVideoRef = useRef<HTMLVideoElement>(null);
+  const localStreamRef = useRef<MediaStream>();
+  const [users, setUsers] = useState<any[]>([]);
 
-  const setVideoTracks = async () => {
+  const getLocalStream = useCallback(async () => {
     try {
-      const stream = await navigator.mediaDevices.getUserMedia({
-        video: true,
+      const localStream = await navigator.mediaDevices.getUserMedia({
         audio: true,
+        video: {
+          width: 240,
+          height: 240,
+        },
       });
-      if (localVideoRef.current) localVideoRef.current.srcObject = stream;
-      if (!(pcRef.current && socketRef.current)) return;
-      stream.getTracks().forEach(track => {
-        if (!pcRef.current) return;
-        pcRef.current.addTrack(track, stream);
-      });
-      pcRef.current.onicecandidate = e => {
-        if (e.candidate) {
-          if (!socketRef.current) return;
-          console.log('onicecandidate');
-          socketRef.current.emit('candidate', e.candidate);
-        }
-      };
-      pcRef.current.oniceconnectionstatechange = e => {
-        console.log(e);
-      };
-      pcRef.current.ontrack = ev => {
-        console.log('add remotetrack success');
-        if (remoteVideoRef.current) {
-          remoteVideoRef.current.srcObject = ev.streams[0];
-        }
-      };
+      localStreamRef.current = localStream;
+      if (localVideoRef.current) localVideoRef.current.srcObject = localStream;
+      if (!socketRef.current) return;
       socketRef.current.emit('join_room', {
         room: '1234',
+        email: 'sample@naver.com',
       });
     } catch (e) {
-      console.error(e);
+      console.log(`getUserMedia error: ${e}`);
     }
-  };
+  }, []);
 
-  const createOffer = async () => {
-    console.log('create offer');
-    if (!(pcRef.current && socketRef.current)) return;
-    try {
-      const sdp = await pcRef.current.createOffer({
-        offerToReceiveAudio: true,
-        offerToReceiveVideo: true,
-      });
-      await pcRef.current.setLocalDescription(new RTCSessionDescription(sdp));
-      socketRef.current.emit('offer', sdp);
-    } catch (e) {
-      console.error(e);
-    }
-  };
+  const createPeerConnection = useCallback(
+    (socketID: string, email: string) => {
+      try {
+        const pc = new RTCPeerConnection(pc_config);
 
-  const createAnswer = async (sdp: RTCSessionDescription) => {
-    if (!(pcRef.current && socketRef.current)) return;
-    try {
-      await pcRef.current.setRemoteDescription(new RTCSessionDescription(sdp));
-      console.log('answer set remote description success');
-      const mySdp = await pcRef.current.createAnswer({
-        offerToReceiveVideo: true,
-        offerToReceiveAudio: true,
-      });
-      console.log('create answer');
-      await pcRef.current.setLocalDescription(new RTCSessionDescription(mySdp));
-      socketRef.current.emit('answer', mySdp);
-    } catch (e) {
-      console.error(e);
-    }
-  };
+        pc.onicecandidate = e => {
+          if (!(socketRef.current && e.candidate)) return;
+          console.log('onicecandidate');
+          socketRef.current.emit('candidate', {
+            candidate: e.candidate,
+            candidateSendID: socketRef.current.id,
+            candidateReceiveID: socketID,
+          });
+        };
+
+        pc.oniceconnectionstatechange = e => {
+          console.log(e);
+        };
+
+        pc.ontrack = e => {
+          console.log('ontrack success');
+          setUsers(oldUsers =>
+            oldUsers
+              .filter(user => user.id !== socketID)
+              .concat({
+                id: socketID,
+                email,
+                stream: e.streams[0],
+              }),
+          );
+        };
+
+        if (localStreamRef.current) {
+          console.log('localstream add');
+          localStreamRef.current.getTracks().forEach(track => {
+            if (!localStreamRef.current) return;
+            pc.addTrack(track, localStreamRef.current);
+          });
+        } else {
+          console.log('no local stream');
+        }
+
+        return pc;
+      } catch (e) {
+        console.error(e);
+        return undefined;
+      }
+    },
+    [],
+  );
 
   useEffect(() => {
     socketRef.current = io.connect(SERVER_URL);
-    pcRef.current = new RTCPeerConnection(pc_config);
+    getLocalStream();
 
-    socketRef.current.on('all_users', (allUsers: Array<{ id: string }>) => {
-      if (allUsers.length > 0) {
-        createOffer();
-      }
-    });
+    socketRef.current.on(
+      'all_users',
+      (allUsers: Array<{ id: string; email: string }>) => {
+        allUsers.forEach(async user => {
+          if (!localStreamRef.current) return;
+          const pc = createPeerConnection(user.id, user.email);
+          if (!(pc && socketRef.current)) return;
+          pcsRef.current = { ...pcsRef.current, [user.id]: pc };
+          try {
+            const localSdp = await pc.createOffer({
+              offerToReceiveAudio: true,
+              offerToReceiveVideo: true,
+            });
+            console.log('create offer success');
+            await pc.setLocalDescription(new RTCSessionDescription(localSdp));
+            socketRef.current.emit('offer', {
+              sdp: localSdp,
+              offerSendID: socketRef.current.id,
+              offerSendEmail: 'offerSendSample@sample.com',
+              offerReceiveID: user.id,
+            });
+          } catch (e) {
+            console.error(e);
+          }
+        });
+      },
+    );
 
-    socketRef.current.on('getOffer', (sdp: RTCSessionDescription) => {
-      //console.log(sdp);
-      console.log('get offer');
-      createAnswer(sdp);
-    });
+    socketRef.current.on(
+      'getOffer',
+      async (data: {
+        sdp: RTCSessionDescription;
+        offerSendID: string;
+        offerSendEmail: string;
+      }) => {
+        const { sdp, offerSendID, offerSendEmail } = data;
+        console.log('get offer');
+        if (!localStreamRef.current) return;
+        const pc = createPeerConnection(offerSendID, offerSendEmail);
+        if (!(pc && socketRef.current)) return;
+        pcsRef.current = { ...pcsRef.current, [offerSendID]: pc };
+        try {
+          await pc.setRemoteDescription(new RTCSessionDescription(sdp));
+          console.log('answer set remote description success');
+          const localSdp = await pc.createAnswer({
+            offerToReceiveVideo: true,
+            offerToReceiveAudio: true,
+          });
+          await pc.setLocalDescription(new RTCSessionDescription(localSdp));
+          socketRef.current.emit('answer', {
+            sdp: localSdp,
+            answerSendID: socketRef.current.id,
+            answerReceiveID: offerSendID,
+          });
+        } catch (e) {
+          console.error(e);
+        }
+      },
+    );
 
-    socketRef.current.on('getAnswer', (sdp: RTCSessionDescription) => {
-      console.log('get answer');
-      if (!pcRef.current) return;
-      pcRef.current.setRemoteDescription(new RTCSessionDescription(sdp));
-      //console.log(sdp);
-    });
+    socketRef.current.on(
+      'getAnswer',
+      (data: { sdp: RTCSessionDescription; answerSendID: string }) => {
+        const { sdp, answerSendID } = data;
+        console.log('get answer');
+        const pc: RTCPeerConnection = pcsRef.current[answerSendID];
+        if (!pc) return;
+        pc.setRemoteDescription(new RTCSessionDescription(sdp));
+      },
+    );
 
     socketRef.current.on(
       'getCandidate',
-      async (candidate: RTCIceCandidateInit) => {
-        if (!pcRef.current) return;
-        await pcRef.current.addIceCandidate(new RTCIceCandidate(candidate));
+      async (data: {
+        candidate: RTCIceCandidateInit;
+        candidateSendID: string;
+      }) => {
+        console.log('get candidate');
+        const pc: RTCPeerConnection = pcsRef.current[data.candidateSendID];
+        if (!pc) return;
+        await pc.addIceCandidate(new RTCIceCandidate(data.candidate));
         console.log('candidate add success');
       },
     );
 
-    setVideoTracks();
+    socketRef.current.on('user_exit', (data: { id: string }) => {
+      if (!pcsRef.current[data.id]) return;
+      pcsRef.current[data.id].close();
+      delete pcsRef.current[data.id];
+      setUsers(oldUsers => oldUsers.filter(user => user.id !== data.id));
+    });
 
     return () => {
       if (socketRef.current) {
         socketRef.current.disconnect();
       }
-      if (pcRef.current) {
-        pcRef.current.close();
-      }
+      users.forEach(user => {
+        if (!pcsRef.current[user.id]) return;
+        pcsRef.current[user.id].close();
+        delete pcsRef.current[user.id];
+      });
     };
-  }, []);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [createPeerConnection, getLocalStream]);
 
   return (
     <Wrapper>
-      <Screen muted ref={localVideoRef} autoPlay />
-      <Screen id="remotevideo" ref={remoteVideoRef} autoPlay />
+      <video muted ref={localVideoRef} autoPlay />
+      <div>
+        {users.map((user, index) => (
+          <Video key={index} stream={user.stream} />
+        ))}
+      </div>
       <Container>
         <TopBar>
           <VideoDetailInfor />
@@ -161,17 +236,16 @@ const Wrapper = styled.div`
   flex-direction: column;
   width: 100%;
   height: 100vh;
-`;
-
-const Screen = styled.video`
-  width: 100%;
-  height: 50%;
-  background-color: #d1d1d1;
-  &:nth-of-type(2) {
-    background-color: #999999;
+  & video {
+    width: 100%;
+    height: 50%;
+    background-color: #d1d1d1;
+    &:nth-of-type(2) {
+      background-color: #999999;
+    }
+    object-fit: cover;
+    transform: scaleX(-1);
   }
-  object-fit: cover;
-  transform: scaleX(-1);
 `;
 
 const Container = styled.div`
